rename file to match hook, add navigation back to semester select
